fixing travis

+++ lib/nutcracker.rb
@@ -49,9 +49,9 @@ module Nutcracker
       return self if attached? or running?
       @pid = ::Process.spawn Nutcracker.executable, *command
       puts "IN HERE"
-      #sleep 10
-      #abort "Nutcracker failed to start" unless running?
-      timeout(5) { sleep 0.1 until running? }
+      sleep 10
+      abort "Nutcracker failed to start" unless running?
+      #timeout(5) { sleep 0.1 until running? }
       Kernel.at_exit { kill if running? }
       self
     end
